refactor(vue): Changed directory structure.
