+++ bindings/vue/src/components/VOnsAlertDialog.vue
@@ -15,8 +15,8 @@
 </template>
 
 <script>
-  import { dialogAPI } from '../internal/mixins/api';
-  import { deriveEvents, deriveMethods, deriveProperties } from '../internal/mixins/derive';
+  import { dialogAPI } from '../mixins/api';
+  import { deriveEvents, deriveMethods, deriveProperties } from '../mixins/derive';
 
   export default {
     mixins: [dialogAPI, deriveEvents, deriveMethods, deriveProperties],
