+++ bindings/vue/src/mixins/common.js
