+++ bindings/vue/src/mixins/derive.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
-import { hyphenate, getClassFromTag, eventToHandler } from '../util';
-import { createMethodsFor, createComputedPropertiesFor } from '../optionsObjectHelper';
+import { hyphenate, getClassFromTag, eventToHandler } from '../internal/util';
+import { createMethodsFor, createComputedPropertiesFor } from '../internal/optionsObjectHelper';
 
 const deriveEvents = {
   beforeCreate() {
